# purplekit/OPENAPI.yaml
openapi: 3.1.0
info:
  title: PurpleKit API
  description: |
    Purple Team Operations Management Platform API
    
    ## Overview
    PurpleKit helps security teams coordinate red team attacks with blue team detection validation,
    track coverage against MITRE ATT&CK, and generate actionable reports.
    
    ## Authentication
    All endpoints (except `/auth/*`) require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Multi-Tenancy
    All data is automatically scoped to the authenticated user's organization.
    You cannot access data from other organizations.
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Report generation: 10 requests/minute
    - ATT&CK sync: 1 request/hour
    
    ## Pagination
    List endpoints support cursor-based pagination:
    - `limit`: Number of items (default: 20, max: 100)
    - `cursor`: Cursor from previous response
    
  version: 1.0.0
  contact:
    name: PurpleKit Support
    email: support@purplekit.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.purplekit.io/v1
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

  # ============================================================================
  # SCHEMAS
  # ============================================================================
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # --------------------------------------------------------------------------
    # Enums
    # --------------------------------------------------------------------------
    UserRole:
      type: string
      enum: [ADMIN, RED_LEAD, BLUE_LEAD, ANALYST, OBSERVER]
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "malcolm@acme.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "securepassword123"
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (expires in 15 minutes)
        refreshToken:
          type: string
          description: Refresh token (expires in 7 days)
        expiresIn:
          type: integer
          description: Access token TTL in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserSummary'
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    # --------------------------------------------------------------------------
    # Users
    # --------------------------------------------------------------------------
    UserSummary:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired refresh token
...